better data gen + real time data transforms
added new classes to model_utils.py to transform the dataset. 

improved dataset generation by only generating portions of the image 
with text in them (to later be rescaled and rotated by the data loader)

a few bugfixes in build_tensor for corner cases.

test_model now handles for the new transformers